feat(pages/game-with-player): add logic to unsubscribe from all events and close socket
Only add to the page of the game-with-player, because most functions require an open socket, so it can only be closed after unmounting this page. Unsubscribe from all events on this page only because each event has multiple subscribers, unsubscribign from any event will unsubscribe in another component of the same event.

feat: add some comments